feat(edit-content) cleanup and reoder store
