+++ core-web/libs/edit-content/src/lib/feature/edit-content/store/features/workflow.feature.ts
@@ -8,29 +8,73 @@ import {
     withState
 } from '@ngrx/signals';
 import { rxMethod } from '@ngrx/signals/rxjs-interop';
-import { pipe } from 'rxjs';
+import { forkJoin, of, pipe } from 'rxjs';
 
 import { HttpErrorResponse } from '@angular/common/http';
 import { computed, inject } from '@angular/core';
+import { Router } from '@angular/router';
+
+import { MessageService, SelectItem } from 'primeng/api';
 
 import { switchMap, tap } from 'rxjs/operators';
 
-import { DotHttpErrorManagerService, DotWorkflowService } from '@dotcms/data-access';
-import { ComponentStatus, DotCMSWorkflow, WorkflowTask } from '@dotcms/dotcms-models';
+import {
+    DotFireActionOptions,
+    DotHttpErrorManagerService,
+    DotMessageService,
+    DotRenderMode,
+    DotWorkflowActionsFireService,
+    DotWorkflowsActionsService,
+    DotWorkflowService
+} from '@dotcms/data-access';
+import {
+    ComponentStatus,
+    DotCMSWorkflow,
+    DotCMSWorkflowAction,
+    WorkflowStep,
+    WorkflowTask
+} from '@dotcms/dotcms-models';
+
+import { ContentState } from './content.feature';
 
 import { EditContentState } from '../edit-content.store';
 
-interface WorkflowState {
+export interface WorkflowState {
+    /** Schemas available for the content type */
+    schemes: {
+        [key: string]: {
+            scheme: DotCMSWorkflow;
+            actions: DotCMSWorkflowAction[];
+            firstStep: WorkflowStep;
+        };
+    };
+
+    /** Current workflow scheme id */
+    currentSchemeId: string | null;
+
+    /** Actions available for the current content */
+    currentContentActions: DotCMSWorkflowAction[];
+
+    /** Current workflow step */
+    currentStep: WorkflowStep | null;
+
+    /** Last workflow task of the contentlet */
+    lastTask: WorkflowTask | null;
+
+    // Nested workflow status
     workflow: {
-        task: WorkflowTask | null;
         status: ComponentStatus;
         error: string | null;
     };
 }
 
 const initialState: WorkflowState = {
+    schemes: {},
+    currentSchemeId: null,
+    currentContentActions: [],
+    currentStep: null,
+    lastTask: null,
     workflow: {
-        task: null,
         status: ComponentStatus.INIT,
         error: null
     }
@@ -45,7 +89,7 @@ const initialState: WorkflowState = {
 export function withWorkflow() {
     return signalStoreFeature(
         {
-            state: type<EditContentState>()
+            state: type<EditContentState & ContentState>()
         },
         withState(initialState),
         withComputed((store) => ({
@@ -65,13 +109,117 @@ export function withWorkflow() {
                 const currentSchemeId = store.currentSchemeId();
 
                 return currentSchemeId ? store.schemes()[currentSchemeId]?.scheme : undefined;
+            }),
+
+            /**
+             * Computed property that determines if workflow action buttons should be shown.
+             * Shows workflow buttons when:
+             * - Content type has only one workflow scheme OR
+             * - Content is existing AND has a selected workflow scheme OR
+             * - Content is new and content type has only one workflow scheme OR
+             * - Content is new and has selected a workflow scheme
+             * Hides workflow buttons when:
+             * - Content is new and has multiple schemes without selection
+             *
+             * @returns {boolean} True if workflow action buttons should be shown, false otherwise
+             */
+            showWorkflowActions: computed(() => {
+                const hasOneScheme = Object.keys(store.schemes()).length === 1;
+                const isExisting = !!store.contentlet();
+                const hasSelectedScheme = !!store.currentSchemeId();
+
+                if (hasOneScheme) {
+                    return true;
+                }
+
+                if (isExisting && hasSelectedScheme) {
+                    return true;
+                }
+
+                if (!isExisting && hasSelectedScheme) {
+                    return true;
+                }
+
+                return false;
+            }),
+
+            /**
+             * Computed property that determines if the workflow selection warning should be shown.
+             * Shows warning when content is new AND no workflow scheme has been selected yet.
+             *
+             * @returns {boolean} True if warning should be shown, false otherwise
+             */
+            showSelectWorkflowWarning: computed(() => {
+                const isNew = !store.contentlet();
+                const hasNoSchemeSelected = !store.currentSchemeId();
+                const hasMultipleSchemas = Object.keys(store.schemes()).length > 1;
+
+                return isNew && hasMultipleSchemas && hasNoSchemeSelected;
+            }),
+
+            /**
+             * Computed property that retrieves the actions for the current workflow scheme.
+             *
+             * @returns {DotCMSWorkflowAction[]} The actions for the current workflow scheme.
+             */
+            getActions: computed(() => {
+                const isNew = !store.contentlet();
+                const currentSchemeId = store.currentSchemeId();
+                const schemes = store.schemes();
+                const currentContentActions = store.currentContentActions();
+
+                // If no scheme is selected, return empty array
+                if (!currentSchemeId || !schemes[currentSchemeId]) {
+                    return [];
+                }
+
+                // For existing content, use specific contentlet actions
+                if (!isNew && currentContentActions.length) {
+                    return currentContentActions;
+                }
+
+                // For new content, use scheme actions
+                return Object.values(schemes[currentSchemeId].actions).sort((a, b) => {
+                    if (a.name === 'Save') return -1;
+                    if (b.name === 'Save') return 1;
+
+                    return a.name.localeCompare(b.name);
+                });
+            }),
+
+            /**
+             * Computed property that transforms the workflow schemes into dropdown options
+             * @returns Array of options with value (scheme id) and label (scheme name)
+             */
+            workflowSchemeOptions: computed<SelectItem[]>(() =>
+                Object.entries(store.schemes()).map(([id, data]) => ({
+                    value: id,
+                    label: data.scheme.name
+                }))
+            ),
+
+            /**
+             * Computed property that retrieves the first step of the current workflow scheme.
+             *
+             * @returns {WorkflowStep} The first step of the current workflow scheme.
+             */
+            getFirstStep: computed(() => {
+                const schemes = store.schemes();
+                const currentSchemeId = store.currentSchemeId();
+
+                return schemes[currentSchemeId]?.firstStep;
             })
         })),
         withMethods(
             (
                 store,
                 dotWorkflowService = inject(DotWorkflowService),
-                dotHttpErrorManagerService = inject(DotHttpErrorManagerService)
+                workflowActionService = inject(DotWorkflowsActionsService),
+                workflowActionsFireService = inject(DotWorkflowActionsFireService),
+                dotHttpErrorManagerService = inject(DotHttpErrorManagerService),
+                messageService = inject(MessageService),
+                dotMessageService = inject(DotMessageService),
+                router = inject(Router)
             ) => ({
                 /**
                  * Get workflow status for an existing contentlet
@@ -120,11 +268,82 @@ export function withWorkflow() {
                     )
                 ),
 
+                /**
+                 * Sets the selected workflow scheme ID in the store.
+                 *
+                 * @param {string} schemeId - The ID of the workflow scheme to be selected.
+                 */
                 setSelectedWorkflow: (schemeId: string) => {
                     patchState(store, {
                         currentSchemeId: schemeId
                     });
-                }
+                },
+
+                /**
+                 * Fires a workflow action and updates the component state accordingly.
+                 *
+                 * This method triggers a sequence of events to fire a workflow action
+                 * and handles the response or error. If the action is successful,
+                 * it navigates to the content view with the updated contentlet and actions.
+                 * In case of an error, it updates the state with an error message.
+                 *
+                 * @param options The options required to fire the workflow action.
+                 */
+                fireWorkflowAction: rxMethod<
+                    DotFireActionOptions<{ [key: string]: string | object }>
+                >(
+                    pipe(
+                        tap(() => patchState(store, { state: ComponentStatus.SAVING })),
+                        switchMap((options) => {
+                            return workflowActionsFireService.fireTo(options).pipe(
+                                tap((contentlet) => {
+                                    if (!contentlet.inode) {
+                                        router.navigate(['/c/content']);
+                                    }
+                                }),
+                                switchMap((contentlet) => {
+                                    return forkJoin({
+                                        currentContentActions: workflowActionService.getByInode(
+                                            contentlet.inode,
+                                            DotRenderMode.EDITING
+                                        ),
+                                        contentlet: of(contentlet)
+                                    });
+                                }),
+                                tapResponse({
+                                    next: ({ contentlet, currentContentActions }) => {
+                                        router.navigate(['/content', contentlet.inode], {
+                                            replaceUrl: true,
+                                            queryParamsHandling: 'preserve'
+                                        });
+
+                                        patchState(store, {
+                                            contentlet,
+                                            currentContentActions,
+                                            state: ComponentStatus.LOADED,
+                                            error: null
+                                        });
+
+                                        messageService.add({
+                                            severity: 'success',
+                                            summary: dotMessageService.get('success'),
+                                            detail: dotMessageService.get(
+                                                'edit.content.success.workflow.message'
+                                            )
+                                        });
+                                    },
+                                    error: (error: HttpErrorResponse) => {
+                                        patchState(store, {
+                                            state: ComponentStatus.LOADED,
+                                            error: 'Error firing workflow action'
+                                        });
+                                        dotHttpErrorManagerService.handle(error);
+                                    }
+                                })
+                            );
+                        })
+                    )
+                )
             })
         )
     );
